fix: remove display_user

+++ back-end/src/bookmarks/repository.rs
@@ -4,7 +4,7 @@ use uuid::Uuid;
 
 use crate::models::{
     bookmark::Bookmark, bookmark_request::CreateBookmarkRequest,
-    update_bookmark_request::UpdateBookmarkRequest,
+    update_bookmark_request::UpdateBookmarkRequest, user,
 };
 
 pub async fn insert_bookmark(pool: &PgPool, bookmark: &Bookmark) -> Result<()> {
@@ -23,21 +23,23 @@ pub async fn insert_bookmark(pool: &PgPool, bookmark: &Bookmark) -> Result<()> {
     Ok(())
 }
 
-pub async fn find_bookmarks(pool: &PgPool) -> Result<Vec<Bookmark>> {
+pub async fn find_bookmarks(pool: &PgPool, user_id: &Uuid) -> Result<Vec<Bookmark>> {
     let bookmarks = sqlx::query_as::<_, Bookmark>(
-        "SELECT id, user_id, title, url, tags, created_at FROM bookmarks",
+        "SELECT id, user_id, title, url, tags, created_at FROM bookmarks WHERE user_id = $1",
     )
+    .bind(user_id)
     .fetch_all(pool)
     .await?;
 
     Ok(bookmarks)
 }
 
-pub async fn find_bookmark(pool: &PgPool, id: &Uuid) -> Result<Option<Bookmark>> {
+pub async fn find_bookmark(pool: &PgPool, id: &Uuid, user_id: &Uuid) -> Result<Option<Bookmark>> {
     let bookmark = sqlx::query_as::<_, Bookmark>(
-        "SELECT id, user_id, title, url, tags, created_at FROM bookmarks WHERE id = $1",
+        "SELECT id, user_id, title, url, tags, created_at FROM bookmarks WHERE id = $1 AND user_id = $2",
     )
     .bind(id)
+    .bind(user_id)
     .fetch_optional(pool)
     .await?;
 
@@ -48,12 +50,13 @@ pub async fn save_bookmark(
     pool: &PgPool,
     id: &Uuid,
     req: &UpdateBookmarkRequest,
+    user_id: &Uuid,
 ) -> Result<Option<Bookmark>> {
     let query = "UPDATE bookmarks
          SET title = COALESCE($1, title),
              url = COALESCE($2, url),
              tags = COALESCE($3, tags)
-         WHERE id = $4
+         WHERE id = $4 AND user_id = $5
          RETURNING id, user_id, title, url, tags, created_at";
 
     let bookmark = sqlx::query_as::<_, Bookmark>(query)
@@ -61,6 +64,7 @@ pub async fn save_bookmark(
         .bind(&req.url)
         .bind(&req.tags)
         .bind(id)
+        .bind(user_id)
         .fetch_optional(pool)
         .await?;
 
@@ -71,12 +75,13 @@ pub async fn replace_bookmark(
     pool: &PgPool,
     id: &Uuid,
     req: &CreateBookmarkRequest,
+    user_id: &Uuid,
 ) -> Result<Option<Bookmark>> {
     let query = "UPDATE bookmarks
          SET title = $1,
              url = $2,
              tags = $3
-         WHERE id = $4
+         WHERE id = $4 AND user_id = $5
          RETURNING id, user_id, title, url, tags, created_at";
 
     let bookmark = sqlx::query_as::<_, Bookmark>(query)
@@ -84,18 +89,19 @@ pub async fn replace_bookmark(
         .bind(&req.url)
         .bind(&req.tags)
         .bind(id)
+        .bind(user_id)
         .fetch_optional(pool)
         .await?;
 
     Ok(bookmark)
 }
 
-pub async fn remove_bookmark(pool: &PgPool, id: &Uuid) -> Result<Option<Bookmark>> {
-    let query =
-        "DELETE FROM bookmarks WHERE id = $1 RETURNING id, user_id, title, url, tags, created_at";
+pub async fn remove_bookmark(pool: &PgPool, id: &Uuid, user_id: &Uuid) -> Result<Option<Bookmark>> {
+    let query = "DELETE FROM bookmarks WHERE id = $1 AND user_id = $2 RETURNING id, user_id, title, url, tags, created_at";
 
     let bookmark = sqlx::query_as::<_, Bookmark>(query)
         .bind(id)
+        .bind(user_id)
         .fetch_optional(pool)
         .await?;
